Rename `FileLinker` to `FileCreationPass` and move
