+++ fuzzyc2cpg/src/test/scala/io/shiftleft/fuzzyc2cpg/CpgTestFixture.scala
@@ -4,8 +4,7 @@ import io.shiftleft.codepropertygraph.Cpg
 import io.shiftleft.fuzzyc2cpg.passes.{AstCreationPass, CMetaDataPass, StubRemovalPass}
 import io.shiftleft.passes.IntervalKeyPool
 import io.shiftleft.semanticcpg.language._
-import io.shiftleft.semanticcpg.passes.CfgCreationPass
-import io.shiftleft.semanticcpg.passes.linking.filecompat.FileLinker
+import io.shiftleft.semanticcpg.passes.{CfgCreationPass, FileCreationPass}
 import io.shiftleft.x2cpg.SourceFiles
 import overflowdb.traversal.TraversalSource
 
@@ -23,7 +22,7 @@ case class CpgTestFixture(projectName: String) {
     new CfgCreationPass(cpg, cfgKeyPool).createAndApply()
   }
   new StubRemovalPass(cpg).createAndApply()
-  new FileLinker(cpg).createAndApply()
+  new FileCreationPass(cpg).createAndApply()
 
   def traversalSource = TraversalSource(cpg.graph)
